+++ src/index.js
@@ -100,10 +100,10 @@ class Squiss extends EventEmitter {
    *    4 days, maximum is 1209600 (14 days).
    * @param {boolean} [opts.autoExtendTimeout=false] If true, the VisibilityTimeout for all in-flight messages will
    *    be automatically extended when there are 5 seconds remaining before the VisibilityTimeout would normally
-   *    expire. It will be extended by VisibilityTimeout + 5 seconds. The VisibilityTimeout used will be the one
+   *    expire. It will be extended by the VisibilityTimeout. The VisibilityTimeout used will be the one
    *    specified in opts.visibilityTimeoutSecs, or the queue's configured VisibilityTimeout if that option is not set.
-   * @param {number} [opts.noExtensionsAfterSecs=0] The age, in seconds, at which a message will no longer have its
-   *    VisibilityTimeout automatically extended if opts.autoExtendTimeout is true.
+   * @param {number} [opts.noExtensionsAfterSecs=43200] The age, in seconds, at which a message will no longer have
+   *    its VisibilityTimeout automatically extended if opts.autoExtendTimeout is true.
    * @param {Object} [opts.queuePolicy] If specified, will be set as the access policy of the queue when
    *    {@link #createQueue} is called. See http://docs.aws.amazon.com/IAM/latest/UserGuide/access_policies.html for
    *    more information.